move makeTestClient to utils
